+++ src/Api/AbstractBuilder.php
@@ -8,11 +8,9 @@ use Evgeek\Moysklad\Api\Segments\AbstractSegmentCommon;
 use Evgeek\Moysklad\Api\Segments\AbstractSegmentNamed;
 use Evgeek\Moysklad\Api\Traits\Actions\DebugTrait;
 use Evgeek\Moysklad\Enums\HttpMethod;
-use Evgeek\Moysklad\Enums\QueryParam;
 use Evgeek\Moysklad\Exceptions\RequestException;
 use Evgeek\Moysklad\Http\ApiClient;
 use Evgeek\Moysklad\Http\Payload;
-use Evgeek\Moysklad\Services\Url;
 use Generator;
 
 abstract class AbstractBuilder
@@ -84,25 +82,5 @@ abstract class AbstractBuilder
     protected function resolveNamedBuilder(string $builderClass): AbstractSegmentNamed
     {
         return new $builderClass($this->api, $this->path, $this->params);
-    }
-
-    protected function setQueryParam(QueryParam|string $queryParam, string|int|float|bool $value): void
-    {
-        $stringQueryParam = strtolower(is_string($queryParam) ? $queryParam : $queryParam->value);
-        $stringValue = Url::convertMixedValueToString($value);
-
-        $separator = QueryParam::getSeparator($stringQueryParam);
-        if ($separator === '') {
-            $this->params[$stringQueryParam] = $stringValue;
-
-            return;
-        }
-
-        if (!array_key_exists($stringQueryParam, $this->params)) {
-            $this->params[$stringQueryParam] = '';
-        }
-        $this->params[$stringQueryParam] .= $this->params[$stringQueryParam] === '' ?
-            $stringValue :
-            $separator . $stringValue;
     }
 }