+++ src/Api/Traits/Params/FilterTrait.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 namespace Evgeek\Moysklad\Api\Traits\Params;
 
 use Evgeek\Moysklad\Enums\FilterSign;
-use Evgeek\Moysklad\Enums\QueryParam;
-use Evgeek\Moysklad\Services\Url;
-use InvalidArgumentException;
+use Evgeek\Moysklad\Services\QueryParams;
 
 trait FilterTrait
 {
@@ -34,90 +32,8 @@ trait FilterTrait
         FilterSign|string|int|float|bool $sign = null,
         string|int|float|bool $value = null
     ): static {
-        if (is_array($key)) {
-            return $this->handleArrayOfFilters($key);
-        }
-
-        [$signString, $valueString] = $this->prepareSignAndValueAsStrings($key, $sign, $value);
-        $filter = $key . $signString . $this->escapeSemicolon($valueString);
-
-        $this->setQueryParam(QueryParam::FILTER, $filter);
+        $this->params = QueryParams::setFilter($this->params, $key, $sign, $value);
 
         return $this;
-    }
-
-    /**
-     * [DEPRECATED]: Use filter() instead.
-     * Filter results with multiple filters. $filters must contain arrays with 3 elements (key, sign and value) or
-     *  only 2 (key and value, '=' will be used as default sign).
-     * <code>
-     * $product = $ms->query()
-     *  ->entity()
-     *  ->product()
-     *  ->filters([
-     *      ['archived', false],
-     *      ['name', '=', 'tangerine'],
-     *      ['code', FilterSign::NEQ, 123],
-     *  ])
-     *  ->get();
-     * </code>
-     *
-     * @see https://dev.moysklad.ru/doc/api/remap/1.2/#mojsklad-json-api-obschie-swedeniq-fil-traciq-wyborki-s-pomosch-u-parametra-filter
-     * @deprecated
-     */
-    public function filters(array $filters): static
-    {
-        return $this->handleArrayOfFilters($filters);
-    }
-
-    private function handleArrayOfFilters(array $filters): static
-    {
-        foreach ($filters as $filter) {
-            if (!is_array($filter)) {
-                throw new InvalidArgumentException('Each filter must be an array');
-            }
-            $this->filter(...$filter);
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return string[]
-     */
-    private function prepareSignAndValueAsStrings(
-        string $key,
-        FilterSign|string|int|float|bool|null $sign,
-        string|int|float|bool|null $value
-    ): array {
-        $prefix = "For filter key '$key': ";
-
-        if ($sign === null) {
-            throw new InvalidArgumentException($prefix . 'sign missed');
-        }
-
-        if ($value === null) {
-            if (is_a($sign, FilterSign::class)) {
-                throw new InvalidArgumentException($prefix . 'with a sign, you must pass the value as the third parameter');
-            }
-
-            /** @var bool|float|int|string $sign */
-            $value = $sign;
-            $sign = $this->defaultSign;
-        } elseif (!is_a($sign, FilterSign::class) && !is_string($sign)) {
-            throw new InvalidArgumentException($prefix . 'with a value, sign must be a string or ' . FilterSign::class);
-        }
-
-        if (is_a($sign, FilterSign::class)) {
-            $sign = $sign->value;
-        }
-        $value = Url::convertMixedValueToString($value);
-
-        return [$sign, $value];
-    }
-
-    private function escapeSemicolon(string $value): string
-    {
-        return str_replace(';', '\;', $value);
     }
 }