+++ src/ApiObjects/Collections/AbstractConcreteCollection.php
@@ -7,9 +7,11 @@ namespace Evgeek\Moysklad\ApiObjects\Collections;
 use Evgeek\Moysklad\ApiObjects\AbstractConcreteApiObject;
 use Evgeek\Moysklad\ApiObjects\Collections\Traits\CrudCollectionTrait;
 use Evgeek\Moysklad\ApiObjects\Collections\Traits\FillMetaCollectionTrait;
+use Evgeek\Moysklad\ApiObjects\Collections\Traits\ParamsCollectionTrait;
 
 abstract class AbstractConcreteCollection extends AbstractConcreteApiObject
 {
     use CrudCollectionTrait;
     use FillMetaCollectionTrait;
+    use ParamsCollectionTrait;
 }