+++ tests/Unit/Api/Traits/Params/ExpandTraitTest.php
@@ -10,10 +10,7 @@ use Evgeek\Moysklad\Enums\HttpMethod;
 use Evgeek\Tests\Unit\Api\Traits\TraitTestCase;
 use InvalidArgumentException;
 
-/**
- * @covers \Evgeek\Moysklad\Api\AbstractBuilder::setQueryParam
- * @covers \Evgeek\Moysklad\Api\Traits\Params\ExpandTrait
- */
+/** @covers \Evgeek\Moysklad\Api\Traits\Params\ExpandTrait */
 class ExpandTraitTest extends TraitTestCase
 {
     public function testSingleStringExpand(): void